refactor permissions request to a condensed form
